requirements.txt
numpy==1.18.5
imutils==0.5.3
scikit-learn==0.23.2
tensorflow==2.3.1
keras==2.4.3
selenium==3.141.0
opencv-python==4.4.0.44
urllib3==1.25.10
matplotlib==3.3.2
Pillow==7.2.0
scikit-image==0.17.2